started 2 and 3 page
started info page and finished 
in the info page added race times, goodies and stuff, and notes. added info that followed those headings.
started the faq page added the questions added questions and answers below that there is a return to top element that works, at the top when you click the question it takes you down to the answer